backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
supabase==2.9.1
python-dotenv==1.0.1
pydantic==2.10.3
